+++ config.def.mk
@@ -1,11 +1,11 @@
 # Modules that don't require external libraries go here
-MODULES=base event noise tone moogladder butter revsc ftbl osc posc3 randi \
+MODULES=base event noise tone moogladder revsc ftbl osc posc3 randi \
 	dcblock tevent metro tenv maygate fofilt biquad tbvcf mode reverse \
 	dtrig rpt dust port vdelay count streson tseq bal drip clip fosc dist \
 	gbuzz rms jitter lpf18 randmt allpass comb phasor tabread \
-	eqfil foo butbr in prop samphold delay switch dmetro random \
+	eqfil foo in prop samphold delay switch dmetro random \
 	pluck atone tin scale pan2 jcrev tenv2 fold decimator inverse mul add sub \
-	div abs panst bar progress pareq
+	div abs panst bar progress pareq butbr butbp buthp butlp
 
 ifndef NO_LIBSNDFILE
 	MODULES += nsmp diskin
@@ -35,7 +35,7 @@ CFLAGS += -I lib/faust
 #MODULES += padsynth
 #CFLAGS += -lfftw3
 
-# If you are on OSX, you may need this 
+# If you are on OSX, you may need this
 # CFLAGS += -I /usr/local/include -L /usr/local/include
 
 # Build dynamic library. Needed for Sporth chugin
