+++ h/butbp.h
@@ -1,0 +1,13 @@
+typedef struct {
+    SPFLOAT sr, freq, bw, istor;
+    SPFLOAT lkf, lkb;
+    SPFLOAT a[8];
+    SPFLOAT pidsr, tpidsr;
+} sp_butbp;
+
+int sp_butbp_create(sp_butbp **p);
+int sp_butbp_destroy(sp_butbp **p);
+int sp_butbp_init(sp_data *sp, sp_butbp *p);
+int sp_butbp_compute(sp_data *sp, sp_butbp *p, SPFLOAT *in, SPFLOAT *out);
+
+int sp_butter_filter(SPFLOAT *in, SPFLOAT *out, SPFLOAT *a);
