+++ h/buthp.h
@@ -1,0 +1,11 @@
+typedef struct  {
+    SPFLOAT sr, freq, istor;
+    SPFLOAT lkf;
+    SPFLOAT a[8];
+    SPFLOAT pidsr;
+} sp_buthp;
+
+int sp_buthp_create(sp_buthp **p);
+int sp_buthp_destroy(sp_buthp **p);
+int sp_buthp_init(sp_data *sp, sp_buthp *p);
+int sp_buthp_compute(sp_data *sp, sp_buthp *p, SPFLOAT *in, SPFLOAT *out);
