--- h/butter.h
@@ -1,37 +1,0 @@
-typedef struct  {
-    SPFLOAT sr, freq, istor;
-    SPFLOAT lkf;
-    SPFLOAT a[8];
-    SPFLOAT pidsr;
-} sp_butlp;
-
-typedef struct  {
-    SPFLOAT sr, freq, istor;
-    SPFLOAT lkf;
-    SPFLOAT a[8];
-    SPFLOAT pidsr;
-} sp_buthp;
-
-typedef struct  {
-    SPFLOAT sr, freq, bw, istor;
-    SPFLOAT lkf, lkb;
-    SPFLOAT a[8];
-    SPFLOAT pidsr, tpidsr;
-} sp_butbp;
-
-int sp_butlp_create(sp_butlp **p);
-int sp_butlp_destroy(sp_butlp **p);
-int sp_butlp_init(sp_data *sp, sp_butlp *p);
-int sp_butlp_compute(sp_data *sp, sp_butlp *p, SPFLOAT *in, SPFLOAT *out);
-
-int sp_buthp_create(sp_buthp **p);
-int sp_buthp_destroy(sp_buthp **p);
-int sp_buthp_init(sp_data *sp, sp_buthp *p);
-int sp_buthp_compute(sp_data *sp, sp_buthp *p, SPFLOAT *in, SPFLOAT *out);
-
-int sp_butbp_create(sp_butbp **p);
-int sp_butbp_destroy(sp_butbp **p);
-int sp_butbp_init(sp_data *sp, sp_butbp *p);
-int sp_butbp_compute(sp_data *sp, sp_butbp *p, SPFLOAT *in, SPFLOAT *out);
-
-int sp_butter_filter(SPFLOAT *in, SPFLOAT *out, SPFLOAT *a);
