+++ modules/butbp.c
@@ -21,102 +21,6 @@
 
 #include "soundpipe.h"
 
-int sp_butlp_create(sp_butlp **p)
-{
-    *p = malloc(sizeof(sp_butlp));
-    return SP_OK;
-}
-
-int sp_butlp_destroy(sp_butlp **p)
-{
-    free(*p);
-    return SP_OK;
-}
-
-int sp_buthp_create(sp_buthp **p)
-{
-    *p = malloc(sizeof(sp_buthp));
-    return SP_OK;
-}
-
-int sp_buthp_destroy(sp_buthp **p)
-{
-    free(*p);
-    return SP_OK;
-}
-
-int sp_butlp_init(sp_data *sp, sp_butlp *p)
-{
-    p->istor = 0.0;
-    p->sr = sp->sr;
-    p->freq = 1000;
-    p->pidsr = M_PI / sp->sr * 1.0;
-    if (p->istor==0.0) {
-        p->a[6] = p->a[7] = 0.0;
-        p->lkf = 0.0;
-    }
-    return SP_OK;
-}
-
-int sp_buthp_init(sp_data *sp, sp_buthp *p)
-{
-    p->istor = 0.0;
-    p->sr = sp->sr;
-    p->freq = 1000;
-    p->pidsr = M_PI / sp->sr * 1.0;
-    if (p->istor==0.0) {
-        p->a[6] = p->a[7] = 0.0;
-        p->lkf = 0.0;
-    }
-    return SP_OK;
-}
-
-int sp_buthp_compute(sp_data *sp, sp_buthp *p, SPFLOAT *in, SPFLOAT *out)
-{
-    if (p->freq <= 0.0)     {
-      *out = 0;
-      return SP_OK;
-    }
-
-    if (p->freq != p->lkf)      {
-      SPFLOAT *a, c;
-      a = p->a;
-      p->lkf = p->freq;
-      c = tan((SPFLOAT)(p->pidsr * p->lkf));
-
-      a[1] = 1.0 / ( 1.0 + ROOT2 * c + c * c);
-      a[2] = -(a[1] + a[1]);
-      a[3] = a[1];
-      a[4] = 2.0 * ( c*c - 1.0) * a[1];
-      a[5] = ( 1.0 - ROOT2 * c + c * c) * a[1];
-    }
-    sp_butter_filter(in, out, p->a);
-    return SP_OK;
-}
-
-int sp_butlp_compute(sp_data *sp, sp_butlp *p, SPFLOAT *in, SPFLOAT *out)
-{
-    if (p->freq <= 0.0){
-      *out = 0;
-      return SP_OK;
-    }
-
-    if (p->freq != p->lkf){
-        SPFLOAT *a, c;
-        a = p->a;
-        p->lkf = p->freq;
-        c = 1.0 / tan((SPFLOAT)(p->pidsr * p->lkf));
-        a[1] = 1.0 / ( 1.0 + ROOT2 * c + c * c);
-        a[2] = a[1] + a[1];
-        a[3] = a[1];
-        a[4] = 2.0 * ( 1.0 - c*c) * a[1];
-        a[5] = ( 1.0 - ROOT2 * c + c * c) * a[1];
-    }
-
-    sp_butter_filter(in, out, p->a);
-    return SP_OK;
-}
-
 int sp_butbp_create(sp_butbp **p)
 {
     *p = malloc(sizeof(sp_butbp));
@@ -177,7 +81,6 @@ int sp_butbp_compute(sp_data *sp, sp_butbp *p, SPFLOAT *in, SPFLOAT *out)
     *out = y;
     return SP_OK;
 }
-
 
 /* Filter loop */
